+++ kubernetes-client/src/test/java/io/fabric8/kubernetes/client/internal/KubernetesResourceUtilTest.java
@@ -20,8 +20,11 @@ import io.fabric8.kubernetes.api.model.ConfigMap;
 import io.fabric8.kubernetes.api.model.ConfigMapBuilder;
 import io.fabric8.kubernetes.api.model.Event;
 import io.fabric8.kubernetes.api.model.EventBuilder;
+import io.fabric8.kubernetes.api.model.Pod;
+import io.fabric8.kubernetes.api.model.PodBuilder;
 import io.fabric8.kubernetes.client.utils.KubernetesResourceUtil;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
@@ -111,5 +114,35 @@ public class KubernetesResourceUtilTest {
     assertEquals("event3", eventList.get(0).getMetadata().getName());
     assertEquals("event2", eventList.get(1).getMetadata().getName());
     assertEquals("event1", eventList.get(2).getMetadata().getName());
+  }
+
+  @Test
+  @DisplayName("Should be able to get resource version")
+  void testGetResourceVersion() {
+    // Given
+    Pod pod = new PodBuilder()
+      .withNewMetadata().withName("test").withResourceVersion("1001").endMetadata()
+      .build();
+
+    // When
+    String resourceVersion = KubernetesResourceUtil.getResourceVersion(pod);
+
+    // Then
+    assertEquals("1001", resourceVersion);
+  }
+
+  @Test
+  @DisplayName("Should be able to update resource version")
+  void testSetResourceVersion() {
+    // Given
+    Pod pod = new PodBuilder()
+      .withNewMetadata().withName("test").withResourceVersion("1001").endMetadata()
+      .build();
+
+    // When
+    KubernetesResourceUtil.setResourceVersion(pod, "1002");
+
+    // Then
+    assertEquals("1002", pod.getMetadata().getResourceVersion());
   }
 }